competition-2/requirements.txt
aiohttp==3.7.4.post0
alembic==1.4.1
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
arviz==0.11.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
bleach==3.3.0
blis==0.7.4
Boruta==0.3
boto3==1.17.46
botocore==1.20.46
catalogue==1.0.0
catboost==0.24.4
certifi==2020.12.5
cffi==1.14.0
cftime==1.4.1
chardet==3.0.4
click==7.1.2
cliff==3.7.0
clikit==0.6.2
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==1.5.0
cmdstanpy==0.9.5
colorama==0.4.4
colorlog==5.0.1
colorlover==0.3.0
conda==4.10.1
conda-build==3.18.11
conda-package-handling==1.7.0
confuse==1.4.0
convertdate==2.3.2
crashtest==0.3.1
cryptography==2.9.2
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
databricks-cli==0.14.3
decorator==4.4.2
defusedxml==0.7.1
docker==5.0.0
entrypoints==0.3
ephem==3.7.7.1
fbprophet==0.7.1
filelock==3.0.12
Flask==1.1.2
funcy==1.15
future==0.18.2
gensim==4.0.1
gevent==21.1.2
gitdb==4.0.7
GitPython==3.1.15
glob2==0.7
glove-python-binary==0.2.0
graphviz==0.16
greenlet==1.0.0
gunicorn==20.1.0
hijri-converter==2.1.1
holidays==0.11.1
htmlmin==0.1.12
httpstan==4.4.2
idna==2.9
ImageHash==4.2.0
imbalanced-learn==0.8.0
importlib-metadata==3.10.0
inotify-simple==1.2.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.13
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kmodes==0.11.0
korean-lunar-calendar==0.2.1
libarchive-c==2.9
lightgbm==3.1.1
llvmlite==0.36.0
LunarCalendar==0.0.9
lz4==3.1.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
matplotlib==3.1.3
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mlflow==1.15.0
mlxtend==0.18.0
multidict==5.1.0
murmurhash==1.0.5
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netCDF4==1.5.6
networkx==2.5.1
nltk==3.6.2
notebook==6.3.0
numba==0.53.1
numexpr==2.7.3
numpy==1.19.5
olefile==0.46
optuna==2.7.0
packaging==20.9
pandas==1.2.4
pandas-profiling==2.11.0
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pastel==0.2.1
patsy==0.5.1
pbr==5.5.1
pexpect==4.8.0
phik==0.11.2
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
plac==1.1.3
plotly==4.14.3
preshed==3.0.5
prettytable==2.1.0
prometheus-client==0.10.0
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.5
protobuf==3.15.7
psutil==5.7.0
ptyprocess==0.6.0
pycaret==2.3.0
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
pyLDAvis==3.3.1
pylev==1.3.0
PyMeeus==0.5.11
PyNaCl==1.4.0
pynndescent==0.5.2
pyod==0.8.7
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.17.3
pysimdjson==3.2.0
PySocks==1.7.1
pystan==3.0.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
querystring-parser==1.2.4
regex==2021.4.4
requests==2.25.1
retrying==1.3.3
ruamel-yaml==0.15.87
s3transfer==0.3.6
sagemaker-training==3.9.0
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
setuptools-git==1.2
six==1.14.0
sklearn==0.0
smart-open==5.0.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.0.1
spacy==2.3.5
SQLAlchemy==1.4.10
sqlparse==0.4.1
srsly==1.0.5
statsmodels==0.12.2
stevedore==3.3.0
tabulate==0.8.9
tangled-up-in-unicode==0.0.7
terminado==0.9.4
testpath==0.4.4
textblob==0.15.3
thinc==7.4.5
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
typing-extensions==3.7.4.3
umap-learn==0.5.1
urllib3==1.25.8
visions==0.6.0
wasabi==0.8.2
wcwidth==0.2.5
webargs==7.0.1
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
xarray==0.17.0
xgboost==1.3.3
yarl==1.6.3
yellowbrick==1.3.post1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
